loader moved from components to common
