+++ src/common/loader/styled.js
